Fixed a couple of issues noticed with the newer burp versions.
With the 2.0 beta versions it includes "beta" with the version. Therefore,
the `new Float(*)` call would fail. Changed it to have some checks for this case.

Also with the changes to burp it appears the saveConfig was deprecated, and therefore
I added a check to see if the version is greater than 1.7 and if so it will use the
saveConfigAsJson call. To parse this I included the GSON library and check the values.
With the new version it will also check if the proxy is running and then grab the values from it.

These fixes should resolve some of the issues seen with #9.